generate db chay ngtraphong va dich vu

+++ SourcesCode/DAO/DICH_VU.cs
@@ -17,6 +17,6 @@ namespace DAO
         public int MaDichVu { get; set; }
         public string TenDichVu { get; set; }
         public string DonViTinh { get; set; }
-        public int DonGia { get; set; }
+        public decimal DonGia { get; set; }
     }
 }
